requirements.txt
certifi==2019.9.11
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dotenv==0.10.3
virtualenv==16.4.3
Werkzeug==0.16.0
